+++ python/lsst/ts/mtdomegui/__init__.py
@@ -32,7 +32,6 @@ else:
 from .application import *
 from .constants import *
 from .control_panel import *
-from .enums import *
 from .main_window import *
 from .model import *
 from .tab import *