+++ doc/version_history.rst
@@ -13,6 +13,7 @@ Version History
 -------------
 
 * Add the dependencies of **ts_mtdomecom** and **ts_config_mttcs**.
+* Remove the **enums.py** and use the enums in **ts_mtdomecom** instead.
 
 .. _lsst.ts.mtdomegui-0.1.0: